Will this get to an end? Am I this stupid? I don't see when perl packages need arrays (or hashes) as parameters or references to them.
'$#' is no longer supported "as of Perl 5.30" - and doesn't work in "5.29.10" neither.

Still only a struggle on the surface!

TODO:
1.) array_humiliation.pl - code multiple arrays in arrays and the dereferencing to an end!

2.) fight perl hashes, hash-references in hashes and dereferencing multiple nested hashes
- 'fat commas' in Hashes: =>
- 'arrow operator': ->

"->" is an infix dereference operator, just as it is in C and C++. If the right side is either a [...], {...}, or a (...) subscript, then the left side must be either a hard or symbolic reference to an array, a hash, or a subroutine respectively. (Or technically speaking, a location capable of holding a hard reference, if it's an array or hash reference being used for assignment.)

Otherwise, the right side is a method name or a simple scalar variable containing either the method name or a subroutine reference, and the left side must be either an object (a blessed reference) or a class name (that is, a package name).