The 'get /' request is handled by an 'action callback'. A callback is a function reference, intended to be called in the future (when a client requests comes in that matches that type of request).
An action is called with one argument, called the controller. The controller is an object that represents our application's interaction with the current transaction. It contains an object representing the transaction, which in turn holds objects for the request and response. One of the object's methods which can be used to generate responses is 'render'.
Mojolicious works with 'the stash': when establishing a route, we can also specify some default values to add to the stash on each request (unless they are changed). These defaults are passed as a hash reference to the route constructor 'get'.